+++ common/context/stories/context-basic.config.mjs
@@ -1,0 +1,14 @@
+export default {
+	sourcePath: '../readme.md',
+	outputName: 'context-basic',
+	story: {
+		title: 'Core/Context',
+		name: 'Context',
+		parameters: {
+			options: {
+				showPanel: false,
+				isToolshown: false
+			}
+		}
+	}
+};