+++ components/mwc-chips/CHANGELOG.md
@@ -1,4 +1,24 @@
-# [0.2.0](https://github.com/material-components/material-components-web-components/compare/v0.17.1...v0.2.0) (2020-12-17)
+# [0.2.0](https://github.com/material-components/material-components-web-components/compare/v1.0.1...v0.2.0) (2021-01-15)
+
+
+
+## [1.0.1](https://github.com/material-components/material-components-web-components/compare/v1.0.0...v1.0.1) (2021-01-15)
+
+
+
+# [1.0.0](https://github.com/material-components/material-components-web-components/compare/v0.19.1...v1.0.0) (2021-01-13)
+
+
+
+## [0.19.1](https://github.com/material-components/material-components-web-components/compare/v0.19.0...v0.19.1) (2021-01-10)
+
+
+
+# [0.19.0](https://github.com/material-components/material-components-web-components/compare/v0.17.2...v0.19.0) (2021-01-07)
+
+
+
+## [0.17.2](https://github.com/material-components/material-components-web-components/compare/v0.17.1...v0.17.2) (2020-12-17)